Move schema CLI next to the function it calls. Move trials enums into trials
