+++ src/boilerdata/main.py
@@ -2,11 +2,10 @@
 
 from types import ModuleType
 
-from typer import Option, Typer
+from typer import Typer
 from typer.main import get_command_name
 
-from boilerdata.configs import write_schema
-from boilerdata.models import Model, model_from_cli
+from boilerdata import configs
 
 
 def add_typer_autoname(app: Typer, module: ModuleType):
@@ -28,18 +27,5 @@ def add_typer_autoname(app: Typer, module: ModuleType):
 
 
 app = Typer()
-for module in []:
+for module in [configs]:
     add_typer_autoname(app, module)
-
-
-@app.command("schema")
-def write_schema_cli(
-    model: Model = Option(
-        ..., help="The Pydantic model.", case_sensitive=False, show_choices=True
-    )
-):
-    """
-    Given a Pydantic model named e.g. "Model", write its JSON schema to
-    "schema/Model.json".
-    """
-    write_schema(f"schema/{model.name}_schema.json", model_from_cli[model])
