+++ tests/test_configs.py
@@ -1,8 +1,11 @@
 from pydantic import BaseModel
 from pytest import mark as m, raises
 import toml
+from typer.testing import CliRunner
 
-from boilerdata import configs
+from boilerdata.configs import app, dump_model, load_config, write_schema
+
+runner = CliRunner()
 
 
 class UserModel(BaseModel):
@@ -40,7 +43,7 @@ SCHEMA_JSON = """\
 @m.parametrize("test_id, file", [("does_not_exist", "file"), ("not_a_file", "")])
 def test_load_config_raises(test_id, file, tmp_path):
     with raises(FileNotFoundError):
-        configs.load_config(tmp_path / file, UserModel)
+        load_config(tmp_path / file, UserModel)
 
 
 def test_load_config_raises_not_toml(tmp_path):
@@ -48,7 +51,7 @@ def test_load_config_raises_not_toml(tmp_path):
     file.touch()
 
     with raises(ValueError):
-        configs.load_config(file, UserModel)
+        load_config(file, UserModel)
 
 
 @m.parametrize(
@@ -61,25 +64,33 @@ def test_load_config_raises_not_toml(tmp_path):
 def test_load_config(test_id, user_model, expected_schema_directive, tmp_path):
     user_model_path = tmp_path / "user_model.toml"
     user_model_path.write_text(user_model)
-    config, schema_directive = configs.load_config(user_model_path, UserModel)
+    config, schema_directive = load_config(user_model_path, UserModel)
     assert toml.load(user_model_path) == config.dict()  # type: ignore
     assert schema_directive == expected_schema_directive
 
 
 def test_dump_model(tmp_path, capfd):
     user_model_path = tmp_path / "test.toml"
-    configs.dump_model(user_model_path, USER_MODEL_INSTANCE)
+    dump_model(user_model_path, USER_MODEL_INSTANCE)
     assert user_model_path.read_text() == USER_MODEL_TOML_NO_SCHEMA
 
 
 def test_write_schema_raises_not_json(tmp_path):
     file = tmp_path / "test.not_json"
     with raises(ValueError):
-        configs.write_schema(file, UserModel)
+        write_schema(file, UserModel)
 
 
 def test_write_schema(tmp_path):
     """Ensure the schema can be written and is up to date."""
     schema_path = tmp_path / "test.json"
-    configs.write_schema(schema_path, UserModel)
+    write_schema(schema_path, UserModel)
     assert schema_path.read_text() == SCHEMA_JSON
+
+
+def test_schema_help():
+    assert runner.invoke(app, ["schema", "--help"]).exit_code == 0
+
+
+def test_schema_trials():
+    assert runner.invoke(app, ["--model", "trials"]).exit_code == 0
